In the BookingQueueSubscriber project, the EndpointUpdatedIntegrationEvent object is looking for a DefenceAdvocate property from the Event. (#589)
The Event firing from the BookingsAPI was firing a EndpointUpdatedIntegrationEvent with a property called DefenceAdvocateContactEmail.
Due to the mismatch of property names the DefenceAdvocate was always null in the BookingQueueSubscriber. This is why the EndPoint Table DefenceAdvocate column was not being updated when editing a hearing.
Updated the EndpointUpdatedIntegrationEvent DefenceAdvocateContactEmail property to DefenceAdvocate.
The UpdateEndpointTests also needed to be updated to use the updated property name.